149325 by sclafani on 2021/05/10 (pair: msakai) Merged change 148856 to Krusty [UDSUP-3] 148856 by jwiele on 2021/03/29 (pair: sweettea) UDSUP-3 Make some data sizes portable.
	Assuming that long int is 64 bits is not always valid. Also,
	while a uint64_t is obviously 64 bits wide, it is not
	necessarily aligned on a 64 bit boundary, so structure padding
	needs to be explicit in order to enforce alignment.

	size_t has to be big enough to hold the size of the largest
	object in memory while off_t has to be big enough to hold the
	size of the largest file system object. These are not
	necessarily the same.

	.../indexLayout.c
	Make structure padding explicit instead of assuming that 64
	  bit integers are 64 bit aligned.

	.../masterIndex005.c
	Declare variables that hold (potentially) large numbers
	  explicitly 64 bits wide, rather than assuming long or size_t
	  is big enough.

	.../fileIORegion.c
	Coerce some (potentially) large quantities to long long and
	  format them with %lld instead of assuming that off_t and
	  size_t are both 64 bits wide and that %ld always requires
	  a 64 bit value.

	.../fileUtils.h
	.../fileUtils.c
	Specify the length to be written as size_t to match what
	  system read and write functions do, as size_t is the
	  portable type to represent the size of any in-memory data.

	.../threadsLinuxUser.c
	Print pthread_t as a pointer. POSIX says that pthread_t
	  doesn't have to be an arithmetic type, but if it is numeric,
	  pointer or struct, it should be printable with %p.

149331 by corwin on 2021/05/10 (pair: sclafani )
Merged 149291 to krusty.
149291 by corwin on 2021/05/10 (pair: sclafani)
Introduce the FORGET idiom.

	This change introduces the FORGET() idiom as a replacement for VDO's
	reference nulling free_ functions. Rather than spelling free_ functions
	as taking a pointer to the pointer to free and having each NULL out the
	passed reference, free_ functions can now take the pointer to free
	directly. The nulling is done by the new FORGET() macro which NULLs out
	a reference to an object and returns a copy of the pointer. Respelled
	free methods can then be invoked as free_foo(FORGET(foo)) rather than
	free_foo(&foo). Furthermore, FORGET can be used when calling FREE()
	directly, or even for any other function call which takes a pointer
	which should not be referenced again once that function returns. For
	example after enqueuing a message.

	.../memoryAlloc.h
	Added the FORGET() macro and the forget() inline function it uses.

	.../buffer.h
	.../buffer.c
	Converted free_buffer() to the new idiom. This is the only free
	  function in UDS which is used by VDO and used the old nulling idiom.

	.../deltaIndex.c
	.../indexConfig.c
	.../indexLayout.c
	.../indexPageMap.c
	.../masterIndex005.c
	.../masterIndex006.c
	Reflected change to free_buffer().

149337 by sclafani on 2021/05/10 (pair: msakai)
Merged changes 148860 and 148939 to Krusty [BZ 1928284] [VDOSTORY-273]
148860 by jwiele on 2021/03/29 (pair: msakai)
BZ 1928284 VDOSTORY-273 Use the map_to_physical_chapter function.

	For a UDS index converted to create space for LVM metadata at
	the front of a VDO, the map_to_physical_chapter function will
	need to be used instead of doing the virtual to physical
	mapping directly because the mapping will be more complicated.

	.../volume.c
	Use map_to_physical_chapter to confirm the virtual to physical
	  chapter mapping instead of doing the calculation directly.

148939 by jwiele on 2021/04/08 (pair: msakai, sclafani, sweettea)
BZ 1928284 VDSTORY-273 Use map_to_physical_chapter instead of calculation.

	Moving a chapter to shrink the volume is going to make
	map_to_physical_chapter more complicated.

	.../geometry.c
	Use map_to_physical_chapter instead of calculating the mapping
	  directly.

149339 by corwin on 2021/05/10 (pair: sclafani)
Reflected change of free_buffer() to new freeing idiom.

	This change introduces the FORGET() macro and the new freeing idiom to
	vdo. Rather than having free_ methods take a pointer to the pointer to
	to free, we use the FORGET() macro, which NULLs out a pointer and
	returns a copy of it. This copy is then passed to the free_ method. In
	this new idiom, the signature of free_foo() changes from
	free_foo(struct foo **fooPtr) to free_foo(struct foo *foo) and no
	longer NULLs fooPtr. The callers change from free_foo(&foo) to
	free_foo(FORGET(foo)). This has the benefit that calls to FREE() or
	other freeing methods which are not in our own code can all be called
	with the same idiom. This has the benefit of further reducing the
	incidence of bugs related to referencing memory after it has been
	freed.

	Furthermore, the FORGET() macro can also be used when enqueuing
	callbacks or in any other situation where a pointer is passed to a
	function and should not then be referenced again once that function
	returns.

	.../superBlockCodec.c
	.../volumeGeometry.c
	Reflected change to free_buffer().

+++ utils/uds/fileIORegion.c
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301, USA. 
  *
- * $Id: //eng/uds-releases/krusty/userLinux/uds/fileIORegion.c#8 $
+ * $Id: //eng/uds-releases/krusty/userLinux/uds/fileIORegion.c#9 $
  */
 
 #include "fileIORegion.h"
@@ -67,9 +67,9 @@ static int validate_io(struct file_io_region *fior,
 
 	if (offset + length > fior->size) {
 		return log_error_strerror(UDS_OUT_OF_RANGE,
-					  "range %zd-%zd not in range 0 to %zu",
-					  offset,
-					  offset + length,
+					  "range %lld-%lld not in range 0 to %zu",
+					  (long long)offset,
+					  (long long)offset + length,
 					  fior->size);
 	}
